# src/api/models/RecipeStepFilter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * amber
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Guid,
    GuidFromJSON,
    GuidFromJSONTyped,
    GuidToJSON,
} from './';

/**
 * 
 * @export
 * @interface RecipeStepFilter
 */
export interface RecipeStepFilter {
    /**
     * 
     * @type {Array<Guid>}
     * @memberof RecipeStepFilter
     */
    recipeIds?: Array<Guid>;
    /**
     * 
     * @type {boolean}
     * @memberof RecipeStepFilter
     */
    includeDeleted?: boolean;
}

export function RecipeStepFilterFromJSON(json: any): RecipeStepFilter {
    return RecipeStepFilterFromJSONTyped(json, false);
}

export function RecipeStepFilterFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecipeStepFilter {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'recipeIds': !exists(json, 'RecipeIds') ? undefined : ((json['RecipeIds'] as Array<any>).map(GuidFromJSON)),
        'includeDeleted': !exists(json, 'includeDeleted') ? undefined : json['includeDeleted'],
    };
}

export function RecipeStepFilterToJSON(value?: RecipeStepFilter | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'RecipeIds': value.recipeIds === undefined ? undefined : ((value.recipeIds as Array<any>).map(GuidToJSON)),
        'includeDeleted': value.includeDeleted,
    };
}