# utils/swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"Guid": {
			"properties": {
				"value": {
					"type": "string"
				}
			},
			"required": [
				"value"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Recipe": {
			"properties": {
				"id": {
					"$ref": "#/definitions/Guid"
				},
				"name": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"requirementIds": {
					"items": {
						"$ref": "#/definitions/Guid"
					},
					"type": "array"
				},
				"readOnly": {
					"properties": {
						"isComplete": {
							"type": "boolean"
						},
						"isScheduled": {
							"type": "boolean"
						},
						"isAssembly": {
							"type": "boolean"
						},
						"endTime": {
							"type": "number",
							"format": "double"
						},
						"startTime": {
							"type": "number",
							"format": "double"
						}
					},
					"required": [
						"isComplete",
						"isScheduled",
						"isAssembly",
						"endTime",
						"startTime"
					],
					"type": "object",
					"x-nullable": true
				}
			},
			"required": [
				"id",
				"name",
				"description",
				"requirementIds",
				"readOnly"
			],
			"type": "object",
			"additionalProperties": true
		},
		"RecipeFilter": {
			"properties": {
				"RecipeStepIds": {
					"items": {
						"$ref": "#/definitions/Guid"
					},
					"type": "array",
					"default": []
				},
				"includeDeleted": {
					"type": "boolean",
					"default": false
				}
			},
			"type": "object",
			"additionalProperties": true
		},
		"FieldDef": {
			"properties": {
				"name": {
					"type": "string"
				},
				"tableID": {
					"type": "number",
					"format": "double"
				},
				"columnID": {
					"type": "number",
					"format": "double"
				},
				"dataTypeID": {
					"type": "number",
					"format": "double"
				},
				"dataTypeSize": {
					"type": "number",
					"format": "double"
				},
				"dataTypeModifier": {
					"type": "number",
					"format": "double"
				},
				"format": {
					"type": "string"
				}
			},
			"required": [
				"name",
				"tableID",
				"columnID",
				"dataTypeID",
				"dataTypeSize",
				"dataTypeModifier",
				"format"
			],
			"type": "object",
			"additionalProperties": true
		},
		"QueryResult_any_": {
			"properties": {
				"command": {
					"type": "string"
				},
				"rowCount": {
					"type": "number",
					"format": "double"
				},
				"oid": {
					"type": "number",
					"format": "double"
				},
				"fields": {
					"items": {
						"$ref": "#/definitions/FieldDef"
					},
					"type": "array"
				},
				"rows": {
					"items": {
						"additionalProperties": true
					},
					"type": "array"
				}
			},
			"required": [
				"command",
				"rowCount",
				"oid",
				"fields",
				"rows"
			],
			"type": "object",
			"additionalProperties": true
		},
		"RecipeStep": {
			"properties": {
				"id": {
					"$ref": "#/definitions/Guid"
				},
				"name": {
					"type": "string"
				},
				"description": {
					"type": "string"
				},
				"recipeRequirementId": {
					"$ref": "#/definitions/Guid"
				},
				"tagId": {
					"$ref": "#/definitions/Guid"
				},
				"recipeId": {
					"$ref": "#/definitions/Guid"
				},
				"resourceId": {
					"$ref": "#/definitions/Guid"
				},
				"duration": {
					"type": "number",
					"format": "double"
				},
				"capacity": {
					"type": "number",
					"format": "double"
				},
				"start": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"id",
				"name",
				"description",
				"recipeRequirementId",
				"tagId",
				"recipeId",
				"resourceId",
				"duration",
				"capacity",
				"start"
			],
			"type": "object",
			"additionalProperties": true,
			"example": {
				"id": {
					"value": "9641905c-8a4b-11eb-8dcd-0242ac130003"
				},
				"name": "Name",
				"description": "Description",
				"recipeRequirementId": {
					"value": "964192e6-8a4b-11eb-8dcd-0242ac130003"
				},
				"tagId": {
					"value": "964193e0-8a4b-11eb-8dcd-0242ac130003"
				},
				"recipeId": {
					"value": "964194bc-8a4b-11eb-8dcd-0242ac130003"
				},
				"resourceId": {
					"value": "9641958e-8a4b-11eb-8dcd-0242ac130003"
				},
				"duration": 0,
				"capacity": 0,
				"start": 0
			}
		},
		"RecipeStepFilter": {
			"properties": {
				"RecipeIds": {
					"items": {
						"$ref": "#/definitions/Guid"
					},
					"type": "array",
					"default": []
				},
				"includeDeleted": {
					"type": "boolean",
					"default": false
				}
			},
			"type": "object",
			"additionalProperties": true
		},
		"Tag": {
			"properties": {
				"id": {
					"$ref": "#/definitions/Guid"
				},
				"name": {
					"type": "string"
				},
				"description": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"name",
				"description"
			],
			"type": "object",
			"additionalProperties": true,
			"example": {
				"id": {
					"value": "0a0796d0-92a2-46c2-bd51-4fc8b63b7b0s"
				},
				"name": "Kettle",
				"description": "Kettle Tag"
			}
		},
		"TagFilter": {
			"properties": {
				"resourceIds": {
					"items": {
						"$ref": "#/definitions/Guid"
					},
					"type": "array",
					"default": []
				},
				"includeDeleted": {
					"type": "boolean",
					"default": false
				}
			},
			"type": "object",
			"additionalProperties": true
		},
		"ResourceReadonly": {
			"properties": {
				"tagList": {
					"type": "string"
				}
			},
			"required": [
				"tagList"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Resource": {
			"properties": {
				"id": {
					"$ref": "#/definitions/Guid"
				},
				"name": {
					"type": "string"
				},
				"tags": {
					"items": {
						"$ref": "#/definitions/Tag"
					},
					"type": "array"
				},
				"readOnly": {
					"$ref": "#/definitions/ResourceReadonly"
				},
				"capacity": {
					"type": "number",
					"format": "double"
				},
				"currentStep": {
					"$ref": "#/definitions/Guid"
				},
				"maintananceRequired": {
					"type": "boolean"
				},
				"active": {
					"type": "boolean"
				}
			},
			"required": [
				"id",
				"name",
				"tags",
				"readOnly",
				"capacity",
				"currentStep",
				"maintananceRequired",
				"active"
			],
			"type": "object",
			"additionalProperties": true
		},
		"ResourceFilter": {
			"properties": {
				"tagIds": {
					"items": {
						"$ref": "#/definitions/Guid"
					},
					"type": "array",
					"default": []
				},
				"includeDeleted": {
					"type": "boolean",
					"default": false
				}
			},
			"type": "object",
			"additionalProperties": true
		}
	},
	"info": {
		"title": "amber",
		"version": "2.0.0",
		"contact": {}
	},
	"paths": {
		"/Recipe/get": {
			"post": {
				"operationId": "GetRecipe",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Recipe"
						}
					}
				},
				"tags": [
					"Recipe"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "recipeId",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Guid"
						}
					}
				]
			}
		},
		"/Recipe/get-by": {
			"post": {
				"operationId": "GetRecipesByFilter",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/Recipe"
							},
							"type": "array"
						}
					}
				},
				"tags": [
					"Recipe"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "filter",
						"required": false,
						"schema": {
							"$ref": "#/definitions/RecipeFilter"
						}
					}
				]
			}
		},
		"/Recipe/delete": {
			"delete": {
				"operationId": "DeleteRecipe",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"tags": [
					"Recipe"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "recipeId",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Guid"
						}
					}
				]
			}
		},
		"/Recipe/update": {
			"put": {
				"operationId": "UpdateOrCreateRecipe",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/QueryResult_any_"
						}
					}
				},
				"tags": [
					"Recipe"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "recipe",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Recipe"
						}
					}
				]
			}
		},
		"/RecipeStep/get": {
			"post": {
				"operationId": "GetRecipeStep",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/RecipeStep"
						}
					}
				},
				"tags": [
					"RecipeStep"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "recipeStepId",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Guid"
						}
					}
				]
			}
		},
		"/RecipeStep/get-by": {
			"post": {
				"operationId": "GetRecipeStepsByFilter",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/RecipeStep"
							},
							"type": "array"
						}
					}
				},
				"tags": [
					"RecipeStep"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "filter",
						"required": false,
						"schema": {
							"$ref": "#/definitions/RecipeStepFilter"
						}
					}
				]
			}
		},
		"/RecipeStep/delete": {
			"delete": {
				"operationId": "DeleteRecipeStep",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"tags": [
					"RecipeStep"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "recipeStepId",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Guid"
						}
					}
				]
			}
		},
		"/RecipeStep/update": {
			"put": {
				"operationId": "UpdateOrCreateRecipeStep",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/QueryResult_any_"
						}
					}
				},
				"tags": [
					"RecipeStep"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "recipeStep",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RecipeStep"
						}
					}
				]
			}
		},
		"/tag/get": {
			"post": {
				"operationId": "GetTag",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Tag"
						}
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "tagId",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Guid"
						}
					}
				]
			}
		},
		"/tag/get-by": {
			"post": {
				"operationId": "GetTagsByFilter",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/Tag"
							},
							"type": "array"
						}
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "filter",
						"required": false,
						"schema": {
							"$ref": "#/definitions/TagFilter"
						}
					}
				]
			}
		},
		"/tag/delete": {
			"delete": {
				"operationId": "DeleteTag",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "tagId",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Guid"
						}
					}
				]
			}
		},
		"/tag/update": {
			"put": {
				"operationId": "UpdateOrCreateTag",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/QueryResult_any_"
						},
						"examples": {
							"application/json": {
								"name": "Mash Tun",
								"description": "Mash Tun"
							}
						}
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "tag",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Tag"
						}
					}
				]
			}
		},
		"/resource/get": {
			"post": {
				"operationId": "GetResource",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Resource"
						}
					}
				},
				"tags": [
					"Resource"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "resourceId",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Guid"
						}
					}
				]
			}
		},
		"/resource/get-by": {
			"post": {
				"operationId": "GetResourcesByFilter",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"$ref": "#/definitions/Resource"
							},
							"type": "array"
						}
					}
				},
				"tags": [
					"Resource"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "filter",
						"required": false,
						"schema": {
							"$ref": "#/definitions/ResourceFilter"
						}
					}
				]
			}
		},
		"/resource/delete": {
			"delete": {
				"operationId": "DeleteResource",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"tags": [
					"Resource"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "resourceId",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Guid"
						}
					}
				]
			}
		},
		"/resource/update": {
			"put": {
				"operationId": "UpdateOrCreateResource",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"additionalProperties": true
						}
					}
				},
				"tags": [
					"Resource"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "resource",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Resource"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {},
	"host": "localhost:3000",
	"schemes": [
		"http"
	]
}